+++ trixnity-core/src/commonMain/kotlin/de/connect2x/trixnity/core/MSCAnnotations.kt
@@ -48,3 +48,19 @@ annotation class MSC2448
     AnnotationTarget.TYPEALIAS
 )
 annotation class MSC4191
+
+/**
+ * @see <a href="https://github.com/matrix-org/matrix-spec-proposals/pull/4143">MSC4143</a>
+ */
+@RequiresOptIn(message = "This API is experimental. It could change in the future without notice.")
+@Retention(AnnotationRetention.BINARY)
+@Target(
+    AnnotationTarget.CLASS,
+    AnnotationTarget.ANNOTATION_CLASS,
+    AnnotationTarget.PROPERTY,
+    AnnotationTarget.FIELD,
+    AnnotationTarget.CONSTRUCTOR,
+    AnnotationTarget.FUNCTION,
+    AnnotationTarget.TYPEALIAS
+)
+annotation class MSC4143